abstracted some things and moved files around
