+++ aries-site/src/components/cards/LinkCard.js
@@ -1,14 +1,14 @@
 import { Box, Card, CardBody, Text } from 'grommet';
 import PropTypes from 'prop-types';
 import Link from 'next/link';
-import { Link as LinkIcon, CirclePlay } from 'grommet-icons';
 
-export const LinkCard = ({ link, title, type }) => {
+export const LinkCard = ({ link, title, icon }) => {
   const Wrapper = Link;
   const wrapperProps = {
     href: link,
     passHref: true,
   };
+  const Icon = icon;
 
   return (
     <Wrapper {...wrapperProps}>
@@ -21,11 +21,7 @@ export const LinkCard = ({ link, title, type }) => {
             background="background-back"
             flex={false}
           >
-            {type === 'guides' ? (
-              <LinkIcon size="large" />
-            ) : (
-              <CirclePlay size="large" />
-            )}
+            <Icon size="large" />
           </Box>
           <Text color="text-strong" weight="bold" size="xlarge">
             {title}
