+++ aries-site/src/examples/foundation/accessibility/index.js
@@ -1,1 +1,2 @@
-export * from './SectionCards';
+export * from './GuideSection';
+export * from './VideoSection';
